+++ src/mpservice/mpserver.py
@@ -1324,23 +1324,11 @@ def _init_server(
     self,
     servlet: Servlet,
     *,
-    main_cpu: int = None,
     capacity: int = 256,
 ):
     self.servlet = servlet
     assert capacity > 0
     self._capacity = capacity
-
-    self._main_cpus = None
-    if main_cpu is not None:
-        # Pin this coordinating thread to the specified CPUs.
-        if isinstance(main_cpu, int):
-            cpus = [main_cpu]
-        else:
-            assert isinstance(main_cpu, list)
-            cpus = main_cpu
-        self._main_cpus = cpus
-        psutil.Process().cpu_affinity(cpus=cpus)
 
 
 def _enter_server(self, gather_args: tuple = None):
@@ -1438,7 +1426,6 @@ class Server:
         self,
         servlet: Servlet,
         *,
-        main_cpu: int = None,
         backlog: int = None,
         capacity: int = 256,
     ):
@@ -1450,10 +1437,6 @@ class Server:
 
             The ``servlet`` has not "started". Its :meth:`~Servlet.start` will be called
             in :meth:`__enter__`.
-
-        main_cpu
-            Specifies the cpu for the "main process",
-            i.e. the process in which this server object resides.
 
         backlog
             .. deprecated:: 0.12.7. Will be removed in 0.14.0. Use `capacity` instead.
@@ -1481,7 +1464,7 @@ class Server:
                 stacklevel=2,
             )
             capacity = backlog
-        _init_server(self, servlet=servlet, capacity=capacity, main_cpu=main_cpu)
+        _init_server(self, servlet=servlet, capacity=capacity)
         self._uid_to_futures = {}
         # Size of this dict is capped at `self._capacity`.
         # A few places need to enforce this size limit.
@@ -1505,9 +1488,6 @@ class Server:
         self.servlet.stop()
         if self._onboard_thread is not None:
             self._onboard_thread.join()
-        if self._main_cpus is not None:
-            psutil.Process().cpu_affinity(cpus=[])
-            # Reset CPU affinity.
 
     def call(self, x, /, *, timeout: int | float = 60, backpressure: bool = True):
         """
@@ -1652,7 +1632,7 @@ class Server:
         *,
         return_x: bool = False,
         return_exceptions: bool = False,
-        timeout: int | float = 600,
+        timeout: int | float = 3600,
     ) -> Iterator:
         """
         Use this method for high-throughput processing of a long stream of
@@ -1791,7 +1771,6 @@ class AsyncServer:
         servlet: Servlet,
         *,
         backlog: int = None,
-        main_cpu: int = None,
         capacity: int = 256,
     ):
         if backlog is not None:
@@ -1801,7 +1780,7 @@ class AsyncServer:
                 stacklevel=2,
             )
             capacity = backlog
-        _init_server(self, servlet=servlet, capacity=capacity, main_cpu=main_cpu)
+        _init_server(self, servlet=servlet, capacity=capacity)
         self._uid_to_futures = {}
         # Size of this dict is capped at `self._capacity`.
         # A few places need to enforce this size limit.
@@ -1826,9 +1805,6 @@ class AsyncServer:
         self.servlet.stop()
         if self._onboard_thread is not None:
             self._onboard_thread.join()
-        if self._main_cpus is not None:
-            psutil.Process().cpu_affinity(cpus=[])
-            # Reset CPU affinity.
 
     async def call(self, x, /, *, timeout: int | float = 60, backpressure: bool = True):
         """
@@ -1952,7 +1928,7 @@ class AsyncServer:
         *,
         return_x: bool = False,
         return_exceptions: bool = False,
-        timeout: int | float = 600,
+        timeout: int | float = 3600,
     ) -> AsyncIterator:
         '''
         Calls to :meth:`stream` and :meth:`call` can happen at the same time
@@ -2035,163 +2011,6 @@ class AsyncServer:
                         yield y
         finally:
             await shutdown(nomoredata, crashed)
-
-
-class StreamServer:
-    '''
-    This class is experimental.
-
-    This class is an alternative to :meth:`Server.stream`.
-    It has a simpler implementation and more focused (narrower but likely enough)
-    use case. This was created in hope of bringing a speedup, but surprisingly
-    it has not demonstrated speed (or throughput) benefits. Until then,
-    there is no plan to deprecate :meth:`Server.stream`.
-    '''
-
-    @final
-    @classmethod
-    def get_mp_context(cls):
-        return MP_SPAWN_CTX
-
-    def __init__(
-        self,
-        servlet: Servlet,
-        *,
-        main_cpu: int = None,
-        capacity: int = 256,
-    ):
-        _init_server(self, servlet=servlet, capacity=capacity, main_cpu=main_cpu)
-
-    @property
-    def capacity(self) -> int:
-        return self._capacity
-
-    def __enter__(self):
-        self._q_in = (
-            _SimpleThreadQueue()
-            if self.servlet.input_queue_type == 'thread'
-            else _SimpleProcessQueue()
-        )
-        self._q_out = (
-            _SimpleThreadQueue()
-            if self.servlet.output_queue_type == 'thread'
-            else _SimpleProcessQueue()
-        )
-        self.servlet.start(self._q_in, self._q_out)
-        return self
-
-    def __exit__(self, *args):
-        self.servlet.stop()
-        if self._main_cpus is not None:
-            psutil.Process().cpu_affinity(cpus=[])
-            # Reset CPU affinity.
-
-    def stream(
-        self,
-        data_stream: Iterable,
-        /,
-        *,
-        return_x: bool = False,
-        return_exceptions: bool = False,
-    ) -> Iterator:
-        """
-        Use this method for high-throughput processing of a long stream of
-        data elements. In theory, this method achieves the throughput upper-bound
-        of the server, as it saturates the pipeline.
-
-        The order of elements in the stream is preserved, i.e.,
-        elements in the output stream correspond to elements
-        in the input stream in the same order.
-
-        This method is NOT thread-safe, that is, there can not be multiple users
-        calling this method concurrently from multiple threads.ResourceWarning
-
-        Parameters
-        ----------
-        data_stream
-            An (possibly unlimited) iterable of input data elements.
-        return_x
-            If ``True``, each output element is a length-two tuple
-            containing the input data and the result.
-            If ``False``, each output element is just the result.
-        return_exceptions
-            If ``True``, any Exception object will be produced in the output stream
-            in place of the would-be regular result.
-            If ``False``, exceptions will be propagated right away, crashing the program.
-        """
-
-        def _enqueue(data_stream, uid_x, nomoredata, stopped):
-            q_in = self._q_in
-            try:
-                for x in data_stream:
-                    if stopped.is_set():
-                        break
-                    uid = id(x)
-                    uid_x.put((uid, x))
-                    q_in.put((uid, x))
-            finally:
-                uid_x.put(nomoredata)
-            # TODO: the workers do not know a stream has concluded.
-            # They may be waiting for some extra time for batching.
-            # This is not a big problem for long jobs, but for short ones
-            # this can be a user experience issue.
-
-        nomoredata = NOMOREDATA
-        uid_x = queue.Queue(self._capacity)
-        stopped = threading.Event()
-
-        worker = Thread(
-            target=_enqueue,
-            args=(data_stream, uid_x, nomoredata, stopped),
-            name=f"{self.__class__.__name__}.stream._enqueue",
-        )
-        worker.start()
-
-        uid_y = {}
-        null = object()
-        q_out = self._q_out
-        n = 0
-        try:
-            while True:
-                z = uid_x.get()
-                # Get one input item from ``uid_x``.
-                # Get its corresponding result, either already in cache ``uid_y``,
-                # or to be obtained out of ``q_out``.
-                if z == nomoredata:
-                    break
-                uid, x = z
-                y = uid_y.pop(uid, null)
-                while y is null:
-                    uid_, y_ = q_out.get()
-                    if uid_ == uid:
-                        y = y_
-                    else:
-                        uid_y[uid_] = y_
-
-                # Exception could happen in the following block,
-                # either by `raise y`, or due to `ExitGenerator`.
-                if isinstance(y, RemoteException):
-                    y = y.exc
-                if isinstance(y, Exception) and not return_exceptions:
-                    raise y
-                if return_x:
-                    yield x, y
-                else:
-                    yield y
-        finally:
-            stopped.set()
-            n = 0  # number of input items that have not been accounted for
-            while z != nomoredata:
-                z = uid_x.get()
-                n += 1
-            n -= 1
-            n -= len(uid_y)
-            # Now need to get count ``n`` results out of ``q_out``.
-            # No need to pair them with input; just discard.
-            for _ in range(n):
-                q_out.get()
-            worker.join()
-
 
 def make_worker(func: Callable[[Any], Any]) -> type[Worker]:
     """