+++ CHANGELOG.md
@@ -8,7 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.13.7] - in progress
 
 - Finetune exit status handling of ``SpawnProcess``.
-
+- ``mpservice.mpserver.StreamServer`` was moved into ``mpservice.experimental.mpserver``.
+- Parameter ``main_cpu`` to a few functions in ``mpservice.mpserver`` was removed.
+- Default value of parameter ``timeout`` to ``mpservice.mpserver.Server.stream`` and ``mpservice.mpserver.AsyncServer.stream``
+  were changed from 600 to 3600, so that user almost never need to specify this parameter.
+  
 
 ## [0.13.6] - 2023-07-18